// src/lib/sanity/types.ts
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};

export type AccessibleImage = {
  _type: "accessibleImage";
  asset?: SanityImageAssetReference;
  media?: unknown;
  hotspot?: SanityImageHotspot;
  crop?: SanityImageCrop;
  alt?: string;
  caption?: string;
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type SETTINGS_QUERY_RESULT = {
  siteName: string | null;
  shortDescription: string | null;
  logo: AccessibleImage | null;
  address: string | null;
  directionsUrl: string | null;
  officeContacts: Array<{
    name: string | null;
    phone: string | null;
  }> | null;
  corporateContacts: Array<{
    name: string | null;
    phone: string | null;
  }> | null;
  whatsapp: string | null;
  socialLinks: Array<{
    label: string | null;
    url: string | null;
  }> | null;
  seo: {
    title: string | null;
    description: string | null;
    image: AccessibleImage | null;
  } | null;
} | null;